Funny, It worked last time
Gameplay improvement commit, it features rebalanced flaps, FAR, DRE and Modular Fuel Tanks patches. It also eliminates a extras as it has been implemented on Tundra itself.

An experimental switch was added on the ship's body, when enabled, Ship's CoM moves up to simulate flaps being folded, it should help with stability but since it's KSP it may as well be useless